test: wait for container to come up in debug test
Two improvements:
1. Right now the test is not waiting for the channel that
runs the goroutine for `appsody debug`, basically not checking
whether it was initiated / ran or not. Put up the container
check loop to make sure the container appears.

2. Name the container uniquely, deriving the name from the stack
so that back-to-back runs / turbulence in the test system do not
confuse the test logic - i.e., read and / or stop a wrong container.