stacks: make mount separator configurable
Right now, `;` is hard-coded to be the mount separator
in the stack-mounts, which inhibits this character from
being used for path names.

Introduce an appsody env APPSODY_MOUNT_SEPARATOR that can
be used by stacks that want to use a different separator.

Partially fixes: https://github.com/appsody/appsody/issues/31