Do not warn if there are no custom stack variables
Custom stack variable are optional, and are unused except for the java
stacks (ATM) but every stack developer will see this inscrutable message
when linting or packaging their stack. Replace it with a not that there
is an unused feature, and a link to the feature docs.

Also, note that the message doesn't use "custom stack variables" (from
the docs) or "templating-data" (the stack yaml key name), making it a
bit hard to understand what it refers to, so clarify the format warning.

- https://appsody.dev/docs/stacks/develop#custom-stack-variables